+++ Engine/Player.h
@@ -29,8 +29,8 @@ private:
 	Vec2 pos;
 	Rect hitbox;
 	Vec2 vel;
-	int shotTimer = 0;
-	static constexpr int refireTime = 15;
+	float shotTimer = 0.0f;
+	static constexpr float refireTime = 15.0f;
 	static constexpr int maxHP = 10;
 	int hp = maxHP;
 };
